+++ src/app/workspace/tests/tester.R
@@ -1,0 +1,4 @@
+library(testthat)
+#test_file(path='tests/test_acmt_network.R', SummaryReporter)
+test_file(path='tests/test_acmt_network.R')
+test_file(path='tests/test_external_data.R')